change template for git grep, hg grep, and bzr grep
Remove case-insensitive option for these templates, since this shouldn't come as a default in my
opinion.
Adapt further the position of the cursor after template insertion. It is right before the '', so
that the user can easily insert then option "-i" and/or go inside '' to specify the pattern for the
search.