prevent *ack* buffer from scrolling in Emacs terminal mode
When visiting or displaying an item in the *ack* buffer the window is
scrolled such that this item is at the top of the window. This annoying
behavior is caused by the fact that ack-mode has compilation-mode as its
parent, and therefore only occurs if Emacs is operated in terminal mode,
or generally if the window has no left fringe [1].
To prevent scrolling the window that displays the *ack* buffer implement
custom functions that mimic the behavior of `compilation-display-error'
and `compilation-goto-error', respectively. This is realized by setting
the variable `compilation-context-lines' locally to an appropriate
integer value (the default value is nil).

Links:

[1] https://www.gnu.org/software/emacs/manual/html_node/emacs/Compilation-Mode.html